src/requirements.txt
pyzbar
Pillow
tldextract
requests
flask
qrcode[pil]
